fix: update source map to reflect .mdx renames and remove guide directory
