+++ .source/browser.ts
@@ -7,6 +7,6 @@ const create = browser<typeof Config, import("fumadocs-mdx/runtime/types").Inter
   }
 }>();
 const browserCollections = {
-  docs: create.doc("docs", {"index.mdx": () => import("../content/index.mdx?collection=docs"), "building-blocks/autobase.md": () => import("../content/building-blocks/autobase.md?collection=docs"), "building-blocks/hyperbee.md": () => import("../content/building-blocks/hyperbee.md?collection=docs"), "building-blocks/hypercore.md": () => import("../content/building-blocks/hypercore.md?collection=docs"), "building-blocks/hyperdht.md": () => import("../content/building-blocks/hyperdht.md?collection=docs"), "building-blocks/hyperdrive.md": () => import("../content/building-blocks/hyperdrive.md?collection=docs"), "building-blocks/hyperswarm.md": () => import("../content/building-blocks/hyperswarm.md?collection=docs"), "guide/creating-a-pear-init-template.md": () => import("../content/guide/creating-a-pear-init-template.md?collection=docs"), "guide/debugging-a-pear-terminal-app.md": () => import("../content/guide/debugging-a-pear-terminal-app.md?collection=docs"), "guide/getting-started.md": () => import("../content/guide/getting-started.md?collection=docs"), "guide/making-a-bare-mobile-app.md": () => import("../content/guide/making-a-bare-mobile-app.md?collection=docs"), "guide/making-a-pear-desktop-app.md": () => import("../content/guide/making-a-pear-desktop-app.md?collection=docs"), "guide/making-a-pear-terminal-app.md": () => import("../content/guide/making-a-pear-terminal-app.md?collection=docs"), "guide/releasing-a-pear-app.md": () => import("../content/guide/releasing-a-pear-app.md?collection=docs"), "guide/sharing-a-pear-app.md": () => import("../content/guide/sharing-a-pear-app.md?collection=docs"), "guide/starting-a-pear-desktop-project.md": () => import("../content/guide/starting-a-pear-desktop-project.md?collection=docs"), "guide/starting-a-pear-terminal-project.md": () => import("../content/guide/starting-a-pear-terminal-project.md?collection=docs"), "helpers/compact-encoding.md": () => import("../content/helpers/compact-encoding.md?collection=docs"), "helpers/corestore.md": () => import("../content/helpers/corestore.md?collection=docs"), "helpers/localdrive.md": () => import("../content/helpers/localdrive.md?collection=docs"), "helpers/mirrordrive.md": () => import("../content/helpers/mirrordrive.md?collection=docs"), "helpers/protomux.md": () => import("../content/helpers/protomux.md?collection=docs"), "helpers/secretstream.md": () => import("../content/helpers/secretstream.md?collection=docs"), "howto/connect-to-many-peers-by-topic-with-hyperswarm.md": () => import("../content/howto/connect-to-many-peers-by-topic-with-hyperswarm.md?collection=docs"), "howto/connect-two-peers-by-key-with-hyperdht.md": () => import("../content/howto/connect-two-peers-by-key-with-hyperdht.md?collection=docs"), "howto/create-a-full-peer-to-peer-filesystem-with-hyperdrive.md": () => import("../content/howto/create-a-full-peer-to-peer-filesystem-with-hyperdrive.md?collection=docs"), "howto/replicate-and-persist-with-hypercore.md": () => import("../content/howto/replicate-and-persist-with-hypercore.md?collection=docs"), "howto/share-append-only-databases-with-hyperbee.md": () => import("../content/howto/share-append-only-databases-with-hyperbee.md?collection=docs"), "howto/work-with-many-hypercores-using-corestore.md": () => import("../content/howto/work-with-many-hypercores-using-corestore.md?collection=docs"), "reference/api.md": () => import("../content/reference/api.md?collection=docs"), "reference/bare-overview.md": () => import("../content/reference/bare-overview.md?collection=docs"), "reference/cli.md": () => import("../content/reference/cli.md?collection=docs"), "reference/configuration.md": () => import("../content/reference/configuration.md?collection=docs"), "reference/faq.md": () => import("../content/reference/faq.md?collection=docs"), "reference/migration.md": () => import("../content/reference/migration.md?collection=docs"), "reference/node-compat.md": () => import("../content/reference/node-compat.md?collection=docs"), "reference/recommended-practices.md": () => import("../content/reference/recommended-practices.md?collection=docs"), "reference/templates.md": () => import("../content/reference/templates.md?collection=docs"), "reference/troubleshooting.md": () => import("../content/reference/troubleshooting.md?collection=docs"), "tools/drives.md": () => import("../content/tools/drives.md?collection=docs"), "tools/hyperbeam.md": () => import("../content/tools/hyperbeam.md?collection=docs"), "tools/hypershell.md": () => import("../content/tools/hypershell.md?collection=docs"), "tools/hyperssh.md": () => import("../content/tools/hyperssh.md?collection=docs"), "tools/hypertele.md": () => import("../content/tools/hypertele.md?collection=docs"), }),
+  docs: create.doc("docs", {"index.mdx": () => import("../content/index.mdx?collection=docs"), "building-blocks/autobase.md": () => import("../content/building-blocks/autobase.md?collection=docs"), "building-blocks/hyperbee.md": () => import("../content/building-blocks/hyperbee.md?collection=docs"), "building-blocks/hypercore.md": () => import("../content/building-blocks/hypercore.md?collection=docs"), "building-blocks/hyperdht.md": () => import("../content/building-blocks/hyperdht.md?collection=docs"), "building-blocks/hyperdrive.md": () => import("../content/building-blocks/hyperdrive.md?collection=docs"), "building-blocks/hyperswarm.md": () => import("../content/building-blocks/hyperswarm.md?collection=docs"), "howto/connect-to-many-peers-by-topic-with-hyperswarm.md": () => import("../content/howto/connect-to-many-peers-by-topic-with-hyperswarm.md?collection=docs"), "howto/connect-two-peers-by-key-with-hyperdht.md": () => import("../content/howto/connect-two-peers-by-key-with-hyperdht.md?collection=docs"), "howto/create-a-full-peer-to-peer-filesystem-with-hyperdrive.md": () => import("../content/howto/create-a-full-peer-to-peer-filesystem-with-hyperdrive.md?collection=docs"), "howto/replicate-and-persist-with-hypercore.md": () => import("../content/howto/replicate-and-persist-with-hypercore.md?collection=docs"), "howto/share-append-only-databases-with-hyperbee.md": () => import("../content/howto/share-append-only-databases-with-hyperbee.md?collection=docs"), "howto/work-with-many-hypercores-using-corestore.md": () => import("../content/howto/work-with-many-hypercores-using-corestore.md?collection=docs"), "helpers/compact-encoding.md": () => import("../content/helpers/compact-encoding.md?collection=docs"), "helpers/corestore.md": () => import("../content/helpers/corestore.md?collection=docs"), "helpers/localdrive.md": () => import("../content/helpers/localdrive.md?collection=docs"), "helpers/mirrordrive.md": () => import("../content/helpers/mirrordrive.md?collection=docs"), "helpers/protomux.md": () => import("../content/helpers/protomux.md?collection=docs"), "helpers/secretstream.md": () => import("../content/helpers/secretstream.md?collection=docs"), "reference/api.mdx": () => import("../content/reference/api.mdx?collection=docs"), "reference/cli.mdx": () => import("../content/reference/cli.mdx?collection=docs"), "reference/configuration.mdx": () => import("../content/reference/configuration.mdx?collection=docs"), "reference/deployment.md": () => import("../content/reference/deployment.md?collection=docs"), "reference/faq.md": () => import("../content/reference/faq.md?collection=docs"), "reference/migration.md": () => import("../content/reference/migration.md?collection=docs"), "reference/recommended-practices.md": () => import("../content/reference/recommended-practices.md?collection=docs"), "reference/runtime.md": () => import("../content/reference/runtime.md?collection=docs"), "reference/troubleshooting.md": () => import("../content/reference/troubleshooting.md?collection=docs"), "tools/drives.md": () => import("../content/tools/drives.md?collection=docs"), "tools/hyperbeam.md": () => import("../content/tools/hyperbeam.md?collection=docs"), "tools/hypershell.md": () => import("../content/tools/hypershell.md?collection=docs"), "tools/hyperssh.md": () => import("../content/tools/hyperssh.md?collection=docs"), "tools/hypertele.md": () => import("../content/tools/hypertele.md?collection=docs"), }),
 };
 export default browserCollections;
