+++ .source/server.ts
@@ -1,41 +1,30 @@
 // @ts-nocheck
-import * as __fd_glob_43 from "../content/tools/hypertele.md?collection=docs"
-import * as __fd_glob_42 from "../content/tools/hyperssh.md?collection=docs"
-import * as __fd_glob_41 from "../content/tools/hypershell.md?collection=docs"
-import * as __fd_glob_40 from "../content/tools/hyperbeam.md?collection=docs"
-import * as __fd_glob_39 from "../content/tools/drives.md?collection=docs"
-import * as __fd_glob_38 from "../content/reference/troubleshooting.md?collection=docs"
-import * as __fd_glob_37 from "../content/reference/templates.md?collection=docs"
-import * as __fd_glob_36 from "../content/reference/recommended-practices.md?collection=docs"
-import * as __fd_glob_35 from "../content/reference/node-compat.md?collection=docs"
-import * as __fd_glob_34 from "../content/reference/migration.md?collection=docs"
-import * as __fd_glob_33 from "../content/reference/faq.md?collection=docs"
-import * as __fd_glob_32 from "../content/reference/configuration.md?collection=docs"
-import * as __fd_glob_31 from "../content/reference/cli.md?collection=docs"
-import * as __fd_glob_30 from "../content/reference/bare-overview.md?collection=docs"
-import * as __fd_glob_29 from "../content/reference/api.md?collection=docs"
-import * as __fd_glob_28 from "../content/howto/work-with-many-hypercores-using-corestore.md?collection=docs"
-import * as __fd_glob_27 from "../content/howto/share-append-only-databases-with-hyperbee.md?collection=docs"
-import * as __fd_glob_26 from "../content/howto/replicate-and-persist-with-hypercore.md?collection=docs"
-import * as __fd_glob_25 from "../content/howto/create-a-full-peer-to-peer-filesystem-with-hyperdrive.md?collection=docs"
-import * as __fd_glob_24 from "../content/howto/connect-two-peers-by-key-with-hyperdht.md?collection=docs"
-import * as __fd_glob_23 from "../content/howto/connect-to-many-peers-by-topic-with-hyperswarm.md?collection=docs"
-import * as __fd_glob_22 from "../content/helpers/secretstream.md?collection=docs"
-import * as __fd_glob_21 from "../content/helpers/protomux.md?collection=docs"
-import * as __fd_glob_20 from "../content/helpers/mirrordrive.md?collection=docs"
-import * as __fd_glob_19 from "../content/helpers/localdrive.md?collection=docs"
-import * as __fd_glob_18 from "../content/helpers/corestore.md?collection=docs"
-import * as __fd_glob_17 from "../content/helpers/compact-encoding.md?collection=docs"
-import * as __fd_glob_16 from "../content/guide/starting-a-pear-terminal-project.md?collection=docs"
-import * as __fd_glob_15 from "../content/guide/starting-a-pear-desktop-project.md?collection=docs"
-import * as __fd_glob_14 from "../content/guide/sharing-a-pear-app.md?collection=docs"
-import * as __fd_glob_13 from "../content/guide/releasing-a-pear-app.md?collection=docs"
-import * as __fd_glob_12 from "../content/guide/making-a-pear-terminal-app.md?collection=docs"
-import * as __fd_glob_11 from "../content/guide/making-a-pear-desktop-app.md?collection=docs"
-import * as __fd_glob_10 from "../content/guide/making-a-bare-mobile-app.md?collection=docs"
-import * as __fd_glob_9 from "../content/guide/getting-started.md?collection=docs"
-import * as __fd_glob_8 from "../content/guide/debugging-a-pear-terminal-app.md?collection=docs"
-import * as __fd_glob_7 from "../content/guide/creating-a-pear-init-template.md?collection=docs"
+import * as __fd_glob_32 from "../content/tools/hypertele.md?collection=docs"
+import * as __fd_glob_31 from "../content/tools/hyperssh.md?collection=docs"
+import * as __fd_glob_30 from "../content/tools/hypershell.md?collection=docs"
+import * as __fd_glob_29 from "../content/tools/hyperbeam.md?collection=docs"
+import * as __fd_glob_28 from "../content/tools/drives.md?collection=docs"
+import * as __fd_glob_27 from "../content/reference/troubleshooting.md?collection=docs"
+import * as __fd_glob_26 from "../content/reference/runtime.md?collection=docs"
+import * as __fd_glob_25 from "../content/reference/recommended-practices.md?collection=docs"
+import * as __fd_glob_24 from "../content/reference/migration.md?collection=docs"
+import * as __fd_glob_23 from "../content/reference/faq.md?collection=docs"
+import * as __fd_glob_22 from "../content/reference/deployment.md?collection=docs"
+import * as __fd_glob_21 from "../content/reference/configuration.mdx?collection=docs"
+import * as __fd_glob_20 from "../content/reference/cli.mdx?collection=docs"
+import * as __fd_glob_19 from "../content/reference/api.mdx?collection=docs"
+import * as __fd_glob_18 from "../content/helpers/secretstream.md?collection=docs"
+import * as __fd_glob_17 from "../content/helpers/protomux.md?collection=docs"
+import * as __fd_glob_16 from "../content/helpers/mirrordrive.md?collection=docs"
+import * as __fd_glob_15 from "../content/helpers/localdrive.md?collection=docs"
+import * as __fd_glob_14 from "../content/helpers/corestore.md?collection=docs"
+import * as __fd_glob_13 from "../content/helpers/compact-encoding.md?collection=docs"
+import * as __fd_glob_12 from "../content/howto/work-with-many-hypercores-using-corestore.md?collection=docs"
+import * as __fd_glob_11 from "../content/howto/share-append-only-databases-with-hyperbee.md?collection=docs"
+import * as __fd_glob_10 from "../content/howto/replicate-and-persist-with-hypercore.md?collection=docs"
+import * as __fd_glob_9 from "../content/howto/create-a-full-peer-to-peer-filesystem-with-hyperdrive.md?collection=docs"
+import * as __fd_glob_8 from "../content/howto/connect-two-peers-by-key-with-hyperdht.md?collection=docs"
+import * as __fd_glob_7 from "../content/howto/connect-to-many-peers-by-topic-with-hyperswarm.md?collection=docs"
 import * as __fd_glob_6 from "../content/building-blocks/hyperswarm.md?collection=docs"
 import * as __fd_glob_5 from "../content/building-blocks/hyperdrive.md?collection=docs"
 import * as __fd_glob_4 from "../content/building-blocks/hyperdht.md?collection=docs"
@@ -51,4 +40,4 @@ const create = server<typeof Config, import("fumadocs-mdx/runtime/types").Intern
   }
 }>({"doc":{"passthroughs":["extractedReferences"]}});
 
-export const docs = await create.docs("docs", "content", {}, {"index.mdx": __fd_glob_0, "building-blocks/autobase.md": __fd_glob_1, "building-blocks/hyperbee.md": __fd_glob_2, "building-blocks/hypercore.md": __fd_glob_3, "building-blocks/hyperdht.md": __fd_glob_4, "building-blocks/hyperdrive.md": __fd_glob_5, "building-blocks/hyperswarm.md": __fd_glob_6, "guide/creating-a-pear-init-template.md": __fd_glob_7, "guide/debugging-a-pear-terminal-app.md": __fd_glob_8, "guide/getting-started.md": __fd_glob_9, "guide/making-a-bare-mobile-app.md": __fd_glob_10, "guide/making-a-pear-desktop-app.md": __fd_glob_11, "guide/making-a-pear-terminal-app.md": __fd_glob_12, "guide/releasing-a-pear-app.md": __fd_glob_13, "guide/sharing-a-pear-app.md": __fd_glob_14, "guide/starting-a-pear-desktop-project.md": __fd_glob_15, "guide/starting-a-pear-terminal-project.md": __fd_glob_16, "helpers/compact-encoding.md": __fd_glob_17, "helpers/corestore.md": __fd_glob_18, "helpers/localdrive.md": __fd_glob_19, "helpers/mirrordrive.md": __fd_glob_20, "helpers/protomux.md": __fd_glob_21, "helpers/secretstream.md": __fd_glob_22, "howto/connect-to-many-peers-by-topic-with-hyperswarm.md": __fd_glob_23, "howto/connect-two-peers-by-key-with-hyperdht.md": __fd_glob_24, "howto/create-a-full-peer-to-peer-filesystem-with-hyperdrive.md": __fd_glob_25, "howto/replicate-and-persist-with-hypercore.md": __fd_glob_26, "howto/share-append-only-databases-with-hyperbee.md": __fd_glob_27, "howto/work-with-many-hypercores-using-corestore.md": __fd_glob_28, "reference/api.md": __fd_glob_29, "reference/bare-overview.md": __fd_glob_30, "reference/cli.md": __fd_glob_31, "reference/configuration.md": __fd_glob_32, "reference/faq.md": __fd_glob_33, "reference/migration.md": __fd_glob_34, "reference/node-compat.md": __fd_glob_35, "reference/recommended-practices.md": __fd_glob_36, "reference/templates.md": __fd_glob_37, "reference/troubleshooting.md": __fd_glob_38, "tools/drives.md": __fd_glob_39, "tools/hyperbeam.md": __fd_glob_40, "tools/hypershell.md": __fd_glob_41, "tools/hyperssh.md": __fd_glob_42, "tools/hypertele.md": __fd_glob_43, });
+export const docs = await create.docs("docs", "content", {}, {"index.mdx": __fd_glob_0, "building-blocks/autobase.md": __fd_glob_1, "building-blocks/hyperbee.md": __fd_glob_2, "building-blocks/hypercore.md": __fd_glob_3, "building-blocks/hyperdht.md": __fd_glob_4, "building-blocks/hyperdrive.md": __fd_glob_5, "building-blocks/hyperswarm.md": __fd_glob_6, "howto/connect-to-many-peers-by-topic-with-hyperswarm.md": __fd_glob_7, "howto/connect-two-peers-by-key-with-hyperdht.md": __fd_glob_8, "howto/create-a-full-peer-to-peer-filesystem-with-hyperdrive.md": __fd_glob_9, "howto/replicate-and-persist-with-hypercore.md": __fd_glob_10, "howto/share-append-only-databases-with-hyperbee.md": __fd_glob_11, "howto/work-with-many-hypercores-using-corestore.md": __fd_glob_12, "helpers/compact-encoding.md": __fd_glob_13, "helpers/corestore.md": __fd_glob_14, "helpers/localdrive.md": __fd_glob_15, "helpers/mirrordrive.md": __fd_glob_16, "helpers/protomux.md": __fd_glob_17, "helpers/secretstream.md": __fd_glob_18, "reference/api.mdx": __fd_glob_19, "reference/cli.mdx": __fd_glob_20, "reference/configuration.mdx": __fd_glob_21, "reference/deployment.md": __fd_glob_22, "reference/faq.md": __fd_glob_23, "reference/migration.md": __fd_glob_24, "reference/recommended-practices.md": __fd_glob_25, "reference/runtime.md": __fd_glob_26, "reference/troubleshooting.md": __fd_glob_27, "tools/drives.md": __fd_glob_28, "tools/hyperbeam.md": __fd_glob_29, "tools/hypershell.md": __fd_glob_30, "tools/hyperssh.md": __fd_glob_31, "tools/hypertele.md": __fd_glob_32, });
